\docType{data}
\name{alldata_eg}
\alias{alldata_eg}
\title{LC-MS metabolomics dataset}
\description{
  alldata_eg is a subset of a cohort study dataset described by De Livera et al. (2015), and
consists of featuredata, sampledata and metabolitedata
}
\examples{
  data(alldata_eg)
  dataview(alldata_eg$featuredata)
  dataview(alldata_eg$metabolitedata)
  dataview(alldata_eg$sampledata)
}
\author{
  Alysha M De Livera, Gavriel Olshansky
}

\references{De Livera, Alysha M, M. Aho-Sysi, Laurent Jacob, J. Gagnon-Bartch, Sandra Castillo, 
J.A. Simpson, and Terence P. Speed. 2015. Statistical methods for handling unwanted 
variation in metabolomics data. \emph{Analytical Chemistry} 87 (7). 
American Chemical Society: 3606-15. 
}

\keyword{datasets}


