\docType{data}
\name{alldataC}
\alias{alldataC}
\title{A designed metabolomics dataset}
\description{
  alldataC is a subsample of the data that consists of 
  featuredata, sampledata and metabolitedata 
  as described by redestig et al 2009
}
\examples{
  data(alldataC)
  dataview(alldataC$featuredata)
  dataview(alldataC$sampledata)
}
\author{
  Alysha M De Livera, Gavriel Olshansky
}
\keyword{datasets}

\references{
  Redestig, H., Fukushima, A., Stenlund, H., Moritz, T., Arita, M., Saito, K.,
  Kusano, M. (2009) Compensation for systematic cross-contribution improves
  normalization of mass spectrometry based metabolomics data. \emph{Analytical
    Chemistry} 81(19): 7974-7980.
  
}
