% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alldatacheck.R
\name{alldatacheck}
\alias{alldatacheck}
\title{Check all data}
\usage{
alldatacheck(featuredata, sampledata = NULL, metabolitedata = NULL)
}
\arguments{
\item{featuredata}{featuredata A data frame in the featuredata format. 
 This is a dataframe with metabolites in columns and samples in rows.
Unique sample names should be provided as row names.
See NormalizeMets Vignette for details.}

\item{sampledata}{A dataframe that contains sample specific information. 
Unique sample names should be provided as row names.}

\item{metabolitedata}{A dataframe that contains metabolite specific information.}
}
\description{
Some preliminary checks of the data formats 
for the NormalizeMets package
}
\author{
Alysha M De Livera, Gavriel Olshansky
}
