% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContrastMatrix.R
\name{ContrastMatrix}
\alias{ContrastMatrix}
\title{Contrast matrix}
\usage{
ContrastMatrix(contrasts, levels)
}
\arguments{
\item{contrasts}{A character vector with specified contrasts.}

\item{levels}{A character vector or a factor with levels in the design
matrix.}
}
\value{
A contrast matrix.
}
\description{
Generates a contrast matrix with specified contrasts.
}
\examples{

    ContrastMatrix(contrasts = c("A-B", "B-C"), levels = c("A", "B", "C", "D"))

}
\seealso{
\code{\link[limma]{makeContrasts}}.
}
\author{
Alysha M De Livera
}
