% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataview.R
\name{dataview}
\alias{dataview}
\title{Data Viewing}
\usage{
dataview(data)
}
\arguments{
\item{data}{a data frame or a data table}
}
\value{
The trimmed data
}
\description{
Prints out a trimmed version of the data matrix
}
\examples{
data("alldata_eg")
featuredata_eg<-alldata_eg$featuredata
dataview(featuredata_eg)

}
\author{
Gavriel Olshanksy and Alysha M De Livera
}
