\docType{data}
\name{Didata}
\alias{Didata}
\title{Direct infusion mass spectrometry metabolomics dataset}
\description{
  Didata (MetaboLights: MTBLS79) consists of featuredata and sampledata as described by 
  Kirwan et al 2014
}
\examples{
  data(Didata)
  dataview(Didata$featuredata)
  dataview(Didata$sampledata)
}
\author{
  Alysha M De Livera, Gavriel Olshansky
}
\keyword{datasets}

\references{
  Kirwan,J.A., Weber,R.J., Broadhurst,D.I. and Viant,M.R. (2014) Direct infusion mass spectrometry metabolomics dataset: a benchmark for data 
  processing and quality control. Sci. Data, 1, 140012.
  
  Haug,K., Salek,R.M., Conesa,P., Hastings,J., de Matos,P., Rijnbeek,M., 
  Mahendraker,T., Williams,M., Neumann,S., Rocca-Serra,P. et al. (2013) 
  MetaboLights--an open-access general-purpose repository for metabolomics studies and associated meta-data. Nucleic Acids Res., 41, D781-786.  
}
