
\docType{data}
\name{featuredata_roots}
\alias{featuredata_roots}
\title{A metabolomics dataset analyzed by GC-TOF/MS.}
\description{
  featuredata_roots consists of featuredata that has been described by Fukushima et al 2013.
}
\examples{
  data(featuredata_roots)
  dataview(featuredata_roots)
}
\author{
  Alysha M De Livera, Gavriel Olshansky
}
\keyword{datasets}

\references{
  Fukushima, A. Gene (2013) 518, 209-214
}



