% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metgroupCheck.r
\name{metgroupCheck}
\alias{metgroupCheck}
\title{Check data compatibility}
\usage{
metgroupCheck(metdata, groupdata)
}
\arguments{
\item{metdata}{A data frame in the met data format. This should have sample
names in the first column to be read as row names and the variables in the
remaining columns. These variables can be metabolites, masses, retention
times, bins, areas or any other metabolomics variables of interest.}

\item{groupdata}{A data frame or a table with optional sample names in the 
first column to be read as row names and group names in the following column.}
}
\description{
Check if the group data matches the metabolite data in size, returns error
message if not compatible
}
\author{
Alysha M De Livera, Gavriel Olshansky
}
