% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplot.r
\name{multiplot}
\alias{multiplot}
\title{Multiple plot function
ggplot objects can be passed in ..., or to plotlist (as a list of ggplot objects)}
\usage{
multiplot(..., plotlist = NULL, file, cols = 1, layout = NULL)
}
\arguments{
\item{...}{polts seperated by ','

If the layout is something like matrix(c(1,2,3,3), nrow=2, byrow=TRUE),
then plot 1 will go in the upper left, 2 will go in the upper right, and
3 will go all the way across the bottom.}

\item{plotlist}{A list of plots.}

\item{file}{A file to save to??..}

\item{cols}{Number of columns in layout}

\item{layout}{A matrix specifying the layout. If present, 'cols' is ignored.}
}
\description{
Multiple plot function
ggplot objects can be passed in ..., or to plotlist (as a list of ggplot objects)
}
\details{
A function adapted from Winston Chang, 'R Graphics Cookbook'.
}
