% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalizeMets-package.R
\docType{package}
\name{NormalizeMets-package}
\alias{NormalizeMets-package}
\alias{NormalizeMets}
\title{NormalizeMets package}
\description{
The NormalizeMets package is a collection of functions for the statistical analysis of
 metabolomics data.

Metabolomics data are inevitably subject to a component of unwanted
variation, due to factors such as batch effects, matrix effects, and confounding
biological variation. This package is a collection of functions designed to implement, 
assess, and choose a suitable normalization method for a given metabolomics study.
}
\details{
Run \code{browseVignettes("NormalizeMets")} or 
\code{vignette("NormalizeMets_vignette", package = "NormalizeMets")} for more details.
}
\author{
Alysha M De Livera, Gavriel Olshansky


Maintainer: Alysha M De Livera <alyshad@unimelb.edu.au>
}
