% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToInputdata.R
\name{ToInputdata}
\alias{ToInputdata}
\title{Convert to Inputdata}
\usage{
ToInputdata(featuredata, groupdata)
}
\arguments{
\item{featuredata}{featuredata A data frame in the featuredata format. 
 This is a dataframe with metabolites in columns and samples in rows.
Unique sample names should be provided as row names.
See NormalizeMets Vignette for details.}

\item{groupdata}{A data frame or a vector with group names.}
}
\value{
\item{inputdata}{A data frame in the input data format. This has
sample names in the first column, group names in the second column, and the
metabolite variables in the remaining columns. }
}
\description{
Returns an inputdata type data frame by combining group information from
groupdata with metabolites information from featuredata
}
\author{
Alysha M De Livera, Gavriel Olshansky
}
