\docType{data}
\name{UVdata}
\alias{UVdata}
\title{A designed metabolomics dataset}
\description{
  UVdata consists of featuredata, sampledata and metabolitedata as described by De Livera et al 2015
}
\examples{
  data(UVdata)
  dataview(UVdata$featuredata)
  dataview(UVdata$metabolitedata)
  dataview(UVdata$sampledata)
}
\author{
  Alysha M De Livera, Gavriel Olshansky
}
\keyword{datasets}

\references{
De Livera, Alysha M De, M. Aho-Sysi, Laurent Jacob, 
J. Gagnon-Bartch, Sandra Castillo, J.A. Simpson, and Terence P. Speed. 
2015. Statistical Methods for Handling Unwanted Variation in 
Metabolomics Data. \emph{Analytical Chemistry} 87 (7). American Chemical Society: 
3606-3615. 
}
